+++ BigExcelCreator/Ranges/CellRange.cs
@@ -1,4 +1,7 @@
-﻿using BigExcelCreator.Extensions;
+﻿// Copyright (c) Federico Seckel.
+// Licensed under the BSD 3-Clause License. See LICENSE file in the project root for full license information.
+
+using BigExcelCreator.Extensions;
 using System;
 using System.Globalization;
 using System.Text;