+++ BigExcelCreator/Styles/StyleElement.cs
@@ -1,4 +1,7 @@
-﻿using DocumentFormat.OpenXml.Spreadsheet;
+﻿// Copyright (c) Federico Seckel.
+// Licensed under the BSD 3-Clause License. See LICENSE file in the project root for full license information.
+
+using DocumentFormat.OpenXml.Spreadsheet;
 
 namespace BigExcelCreator.Styles
 {