+++ BigExcelCreator/Styles/StyleList.cs
@@ -1,4 +1,7 @@
-﻿using DocumentFormat.OpenXml;
+﻿// Copyright (c) Federico Seckel.
+// Licensed under the BSD 3-Clause License. See LICENSE file in the project root for full license information.
+
+using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Spreadsheet;
 using System;
 using System.Collections.Generic;